requirements.txt
streamlit>=1.28.0
streamlit-authenticator>=0.3.1
bcrypt>=4.0.0
astrapy>=1.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-groq>=0.0.1
extra-streamlit-components>=0.1.60

